+++ NEWS.md
@@ -1,5 +1,9 @@
 # tidycmprsk (development version)
 
+* Now returning the the number at risk, cumulative number of events, and the cumulative number of censored observations in `tidy.cuminc()`.
+
+* Updated the default value to `tidy.cuminc(conf.int = TRUE)` from `FALSE`
+
 * Update the `crr()` print method for more clarity.
 
 # tidycmprsk 0.1.0